Convert export-related code to plugins per format
Patch adds internal infrastructure for export plugins, and moves
existing code into such plugins.

Export plugins are *not* the same kind of plugins tmt uses for steps.
Step plugins are way too connected to how steps work - it will be useful
to share some of the basic concepts like loading and "how" registries,
but the change would be too complex. It will happen in the future,
though.

The overlap is not that big anyway, both kinds are imported by tmt, both
register their functionality, and that's all.

Export plugins may implement methods for exporting various levels of
metadata, from fmf ids to tests, plans and stories. Existing code has
been refactored to 1. move export code to plugins, and 2. turn in-class
export methods to use the plugins.

There are some intersections between some of the plugins, e.g. Polarion
export uses some Nitrate functionality, it might be possible to clean it
up a bit, but that can wait for future patches. The primary goal was to
bootstrap export plugins and split existing code.

One particular issue has been fixed by introducing the concept of
exporting "collections" of objects: when exported one by one to YAML,
objects were exported as mappings, one by one, without document
separators between them. This made it hard for tools like `yq` to digest
and process them correctly.

The biggest blocks of changed lines are mostly moves. Nitrate and
Polarion code was moved into their own files, but not really changed,
with exception of imports and added export plugin classes.
