+++ tests/integration/__init__.py
@@ -6,7 +6,7 @@ from requre.cassette import StorageKeysInspectSimple
 from requre.helpers.guess_object import Guess
 from requre.helpers.requests_response import RequestResponseHandling
 
-import tmt.export
+import tmt.utils
 
 nitrate.set_cache_level(nitrate.CACHE_NONE)
 
@@ -23,7 +23,7 @@ sessions.Session.send = RequestResponseHandling.decorator(
     item_list=[1])(
         sessions.Session.send)
 
-tmt.export.check_git_url = Guess.decorator_plain()(tmt.export.check_git_url)
+tmt.utils.check_git_url = Guess.decorator_plain()(tmt.utils.check_git_url)
 
 
 # use storage simple strategy to avoid use full stack info for keys
